Introduce commander to act as the "front controller" to different features
