+++ package.json
@@ -4,8 +4,9 @@
   "description": "Command line utility to make configuring and setting up Gitevents easier.",
   "main": "index.js",
   "dependencies": {
-    "inquirer": "^0.11.2",
-    "contributor": "^0.1.25"
+    "commander": "^2.9.0",
+    "contributor": "^0.1.25",
+    "inquirer": "^0.11.2"
   },
   "devDependencies": {
     "tap-spec": "^4.1.1",
